Adds size validation in Mac Address and changes a test oid
Adds a Mac Address size validation of 17. This excludes the possibility of Go's native ParserMAC to identify Mac of type EUI-64 and 20-octet. But it still identifies the EUI-48 standard that is not the Mac Address of network devices.
It changes oids to test, this change was random to try to identify other types of data provided by SNMP.